feat: align process, SDK step, and flow logic
Includes significant refactoring to old 'top-heavy' config schema to allow for easier extension in future and separate it from the out-of-the-box configuration in the <configdatastorage> element.

Services classes have also been updated to be aligned with the primary entity that they target (e.g. DocumentTemplateDeploymentService is preferred over WordTemplateDeploymentService). This avoids arbitrary and ambiguous separation of the service classes.

Removed redundant <flowstodeactivate> element as flows can fall under the existing <processes> configuration.

Moved process and SDK step state setting to run before post-deploy data import. If processes must be inactive only during the import, this can be configured in the data import configuration file outside of the template.